Organize code & Add README
