+++ template_ws/src/minimal_pkg/include/minimal_pkg/subscriber.hpp
@@ -37,11 +37,3 @@ private:
   }
   rclcpp::Subscription<std_msgs::msg::String>::SharedPtr subscription_;
 };
-
-int main(int argc, char * argv[])
-{
-  rclcpp::init(argc, argv);
-  rclcpp::spin(std::make_shared<MinimalSubscriber>());
-  rclcpp::shutdown();
-  return 0;
-}
